requirements.txt
asgiref==3.2.3
astroid==2.3.3
azure-functions==1.2.0
azure-functions-wsgi-adapter==0.2
Django==3.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.4
ptvsd==4.3.2
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
wrapt==1.11.2
